LayoutTests/platform/mac/media/mediacapabilities/hevc-decodingInfo-hdr-expected.txt

RUN(internals.settings.setMediaCapabilitiesExtensionsEnabled(true))
RUN(promise = navigator.mediaCapabilities.decodingInfo({ type: 'media-source', video: { contentType: 'video/mp4; codecs="hvc1.2.4.H153.B0"', height: 720, bitrate: 800000, width: 1280, framerate: 24.5 }});)
Promise resolved OK
EXPECTED (info.supported == 'true') OK
EXPECTED (info.supportedConfiguration.video.hdrMetadataType == 'null') OK
RUN(promise = navigator.mediaCapabilities.decodingInfo({ type: 'media-source', video: { contentType: 'video/mp4; codecs="hvc1.2.4.H153.B0"', height: 720, bitrate: 1000, width: 1280, framerate: 24.5, hdrMetadataType: 'smpteSt2086', transferFunction: 'pq' }});)
Promise resolved OK
EXPECTED (info.supported == 'true') OK
EXPECTED (info.supportedConfiguration.video.hdrMetadataType == 'smpteSt2086') OK
RUN(promise = navigator.mediaCapabilities.decodingInfo({ type: 'media-source', video: { contentType: 'video/mp4; codecs="avc1.640028"', height: 720, bitrate: 1000, width: 1280, framerate: 24.5, hdrMetadataType: 'smpteSt2086', transferFunction: 'pq' }});)
Promise resolved OK
EXPECTED (info.supported == 'false') OK
EXPECTED (info.supportedConfiguration.video.hdrMetadataType == 'smpteSt2086') OK
END OF TEST
